Refactor getImageSize() into ImageUtils Service
Changes:
 - Refactor Card component to use ImageUtils.getImageSize()
 - Refactor CardListItem component to use ImageUtils.getImageSize()
