+++ App/Services/ImageUtils.js
@@ -1,0 +1,13 @@
+export const getImageSize = (imageUri) => {
+  return new Promise((resolve, reject) => {
+    Image.getSize(imageUri,
+      (width, height) => {
+        const maxWidth = Metrics.screenWidth - (Metrics.unit * 4)
+        const size = (maxWidth < width)
+          ? { width: maxWidth, height: height * (maxWidth / width) }
+          : { width, height }
+        resolve(size)
+      },
+      reject)
+  })
+}
